Fix steering vector calculation
Steering vectors depend on the frequency of the signal-of-interest, but
not on the central frequency of a DAQ as it is now. So lets compensate
for the discrepancy. Also, for convenience lets set VFOs frequencies to
the central one on receiver parameters update. Finally, we should use
radius of UCA for steering vectors calculation, but since
b43e5e8b8b94b0918a2e7b1e1ae9a33ed3318a9a we use inter element spacing
instead.